+++ src/viewer/icons/index.ts
@@ -1,3 +1,4 @@
+import CloseIcon from "./close.svg";
 import CollapseIcon from "./collapse.svg";
 import CopyDeckIcon from "./copy_deck.svg";
 import ExpandIcon from "./expand.svg";
@@ -8,6 +9,7 @@ import UnpinIcon from "./unpin.svg";
 export {
 	CollapseIcon,
 	ExpandIcon,
+	CloseIcon,
 	PinIcon,
 	UnpinIcon,
 	CopyDeckIcon,